refactor(server): extract OpenAI SSE assembler
Move transcript.delta / done payloads and OpenAIStreamAssembler
into openai/sse.rs. Public openai::* paths are unchanged.
